fix: fail closed in firewall base url matching
