+++ turbo/packages/connectors/src/firewall-expander.ts
@@ -4,6 +4,7 @@ import {
   validateAuthBaseUrl,
   validateBaseUrl,
 } from "./firewall-types";
+import { hasRawWhitespace, hasUnsafeUrlCodepoint } from "./firewall-url-utils";
 import { parseSegment, splitPathSegments } from "./segment-parser";
 import { fetchFirewallConfig, type FetchFn } from "./firewall-loader";
 
@@ -21,61 +22,6 @@ const VALID_RULE_METHODS = new Set([
   "OPTIONS",
   "ANY",
 ]);
-const ASCII_CONTROL_MAX = 0x20;
-const ASCII_DELETE = 0x7f;
-const UNICODE_HIGH_SURROGATE_MIN = 0xd800;
-const UNICODE_HIGH_SURROGATE_MAX = 0xdbff;
-const UNICODE_LOW_SURROGATE_MIN = 0xdc00;
-const UNICODE_LOW_SURROGATE_MAX = 0xdfff;
-
-function hasRawWhitespace(value: string): boolean {
-  for (let i = 0; i < value.length; i += 1) {
-    const char = value[i]!;
-    if (
-      char === " " ||
-      char === "\t" ||
-      char === "\n" ||
-      char === "\r" ||
-      char === "\f" ||
-      char === "\v"
-    ) {
-      return true;
-    }
-  }
-  return false;
-}
-
-function hasUnsafeUrlCodepoint(value: string): boolean {
-  for (let i = 0; i < value.length; i += 1) {
-    const codeUnit = value.charCodeAt(i);
-    if (codeUnit < ASCII_CONTROL_MAX || codeUnit === ASCII_DELETE) {
-      return true;
-    }
-    if (
-      UNICODE_HIGH_SURROGATE_MIN <= codeUnit &&
-      codeUnit <= UNICODE_HIGH_SURROGATE_MAX
-    ) {
-      const nextCodeUnit = value.charCodeAt(i + 1);
-      if (
-        !(
-          UNICODE_LOW_SURROGATE_MIN <= nextCodeUnit &&
-          nextCodeUnit <= UNICODE_LOW_SURROGATE_MAX
-        )
-      ) {
-        return true;
-      }
-      i += 1;
-      continue;
-    }
-    if (
-      UNICODE_LOW_SURROGATE_MIN <= codeUnit &&
-      codeUnit <= UNICODE_LOW_SURROGATE_MAX
-    ) {
-      return true;
-    }
-  }
-  return false;
-}
 
 function validatePathSegments(
   path: string,
